multithreading debugging work
worked on debugging multithreading code; currently has serial reading running smoothly in it's own thread with the graph function at least able to receive the data in serial_record. More work needs to be done to ensure that serial_record is thread safe and that it isn't being filled with gibberish as fast_decoed Immediately detected a misalignment. Further improvement on misalignment code needed, as running it crashed the graph thread and the serial read thread.